+++ app/src/main/java/com/nmrc/datastructure/components/tree_screen/domain/BinaryTree.kt
@@ -4,9 +4,9 @@ abstract class BinaryTree {
     open var size: Int = 0
         protected set
 
-    abstract fun insert(value: Int)
-    abstract fun remove(value: Int)
-    abstract fun contains(value: Int): Boolean
+    abstract fun insert(value: Float)
+    abstract fun remove(value: Float)
+    abstract fun contains(value: Float): Boolean
     abstract fun returnComposableData(): List<NodeComposableData>
     abstract fun asBalancedTree(): BinaryNodeTree
     abstract fun heapify(isMin: Boolean): HeapTree
